\name{TinnRcom-package} 
\alias{TinnRcom-package} 
\docType{package} 
 
\title{ 
  Resources of Tinn-R GUI/Editor for R Environment 
} 
 
\description{ 
  Implements a set of customized functions, inspired from \pkg{svIDE}, \pkg{svMisc} 
  and \pkg{svIO} packages of Philippe Grosjean and co-authors as soon as \pkg{vimcom}
  of Jakson A. de Aquino, required to Tinn-R GUI/Editor for R environment. 
} 
 
\author{
  José Cláudio Faria \email{joseclaudio.faria@gmail.com}\cr
}

\references{ 
  \url{URL: http://www.sciviews.org/SciViews-R}\cr
  \url{http://www.lepem.ufc.br/aquino.php}  
} 
 
\keyword{package} 
