\name{trPaths}
\alias{trPaths}

\title{
  Short paths to Tinn-R Editor
}

\description{
  The ocult object .trPaths stores a vector of characters related to the use of Tinn-R Editor.
}

\value{
  A vector of characters.
}

\author{
  José Cláudio Faria \email{joseclaudio.faria@gmail.com}\cr
}

\keyword{utilities}
