\name{trPaths}
\alias{trPaths}

\title{
  Short paths to Tinn-R Editor
}

\description{
  \code{trPaths()} store a vector of characters related to the use of Tinn-R Editor.
}

\value{
  A vector of characters.
}

\author{
  Original authors: Philippe Grosjean \email{phgrosjean@sciviews.org}\cr
                    Jakson A. de Aquino \email{jalvesaq@gmail.com}\cr
  Adaptations: José Cláudio Faria \email{joseclaudio.faria@gmail.com}\cr
}

\keyword{ utilities }
